Rename loader to snake case
