+++ Sprocket/Graphics/Rendering/Scene3DRenderer.cpp
@@ -4,7 +4,7 @@
 #include <Sprocket/Graphics/buffer.h>
 #include <Sprocket/Graphics/open_gl.h>
 #include <Sprocket/Graphics/render_context.h>
-#include <Sprocket/Scene/Camera.h>
+#include <Sprocket/Scene/camera.h>
 #include <Sprocket/Utility/Hashing.h>
 #include <Sprocket/Utility/Maths.h>
