+++ Sprocket/Scene/loader.dmx.py
