+++ Sprocket/Scene/loader.h
